% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePC.R
\name{annotatePC}
\alias{annotatePC}
\title{Annotate top PCs from the dataset}
\usage{
annotatePC(PCnum, val_all, PCAmodel, n = 5, scoreCutoff = 0.5, simplify = TRUE)
}
\arguments{
\item{PCnum}{PC number of your dataset you want to get the annotation results. It
should be an integer value between 1 and 8.}

\item{val_all}{The output from \code{\link{validate}}}

\item{PCAmodel}{The PCAmodel used to generate the argument, \code{val}}

\item{n}{An integer. Default is 5. The number of the top enriched pathaways to
print out. If there are fewer than n pathways passed the cutoff, it will print out \code{NA}.}

\item{scoreCutoff}{A numeric value for the minimum correlation. Default is 0.5.}

\item{simplify}{A logical. Under default (\code{TRUE}), the output will be a
data frame with the number of column same as the length of \code{PCnum} argument,
and the number of row same as the \code{n} argument. If it is set to \code{FALSE},
the output will be a list with the length of \code{PCnum} argument, where each
element is a data frame containing detailed GSEA output of enriched pathways.}
}
\value{
A data frame of a list based on the \code{simplify} argument. Check the
output detail above.
}
\description{
This function finds the PCcluster scored highest with the top PCs of the dataset,
including PCclusters with the negative average silhouette.
}
