% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateScore.R
\name{calculateScore}
\alias{calculateScore}
\title{Calculate the validation score for a new dataset}
\usage{
calculateScore(dataset, PCAmodel)
}
\arguments{
\item{dataset}{A gene expression matrix, where genes are in rows and rownames
are in 'symbol' format. It can be SummarizedExperiment, ExpressionSet, or matrix objects.}

\item{PCAmodel}{PCAGenomicSignatures object. Output from \code{buildAvgLoading}
function, a matrix of avaerage loadings, can be directly provided.}
}
\value{
A list containing the score matrices for input datasets. Scores are
assigned to each sample (row) for each cluster (column).
}
\description{
Calculate the validation score for a new dataset
}
