% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonGene.R
\name{commonGene}
\alias{commonGene}
\title{Common genes from datasets}
\usage{
commonGene(dat, loaded = TRUE)
}
\arguments{
\item{dat}{A character vector containing the name of datasets loaded in memory.
Currently, the loaded dataset format should be in ExpressionSet class. You
can also provide a list of genes you want to merge into a common gene list.}

\item{loaded}{The default condition (\code{TRUE}) assumes that the dataset (e.g.
expression matrix with genes as row names) is loaded. Set this argument as
\code{FALSE} if you provide a list of genes in the argument \code{dat}.}
}
\value{
A character vector with the common genes of the datasets.
}
\description{
Common genes from datasets
}
\note{
This function is for CRC datasets.
}
