% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAGenomicSignatures-class.R
\name{.PCclusterName}
\alias{.PCclusterName}
\title{Formatting PCcluster name}
\usage{
.PCclusterName(x, ...)
}
\arguments{
\item{x}{PCAGenomicSignature-object}

\item{...}{Additional arguments for supporting functions.}
}
\description{
Keep the name with 'k + cluster number + number of PCs + number of unique studies'
info during the model construction to make it easy to keep track of them, but at the
PCAGenomicSignatures-class object building step, covert them into a short version,
mPC (mean of PCs).
}
