% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{droplist}
\alias{droplist}
\title{MeSH terms to be excluded in drawWordcloud function}
\format{
A character vector containing MeSH terms to be excluded.
}
\usage{
droplist
}
\description{
MeSH terms to be excluded in drawWordcloud function
}
\keyword{datasets}
