% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPC.R
\name{extractPC}
\alias{extractPC}
\title{PCA on gene expression profile}
\usage{
extractPC(x)
}
\arguments{
\item{x}{a numeric or complex matrix (or data frame) which provides the gene
expression data for the principal components analysis. Genes in the rows and
samples in the columns.}
}
\value{
A \code{\link[stats]{prcomp}} object.
}
\description{
Performs a principal components analysis on the given data matrix and returns
the results as an object of class \code{\link[stats]{prcomp}}.
}
\examples{
m = matrix(rnorm(100),ncol=5)
extractPC(m)

}
\seealso{
\code{\link[stats]{prcomp}}
}
