% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filterList}
\alias{filterList}
\title{RAVs that will output with quanlity-control messages}
\format{
A named list with four elements - "Cluster_Size_filter",
"GSEA_C2_filter", "GSEA_PLIERpriors_filter", and "Redundancy_filter".
}
\usage{
filterList
}
\description{
RAVs that will output with quanlity-control messages
}
\author{
Sehyun Oh \email{shbrief@gmail.com}
}
\keyword{datasets}
