% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSignature.R
\name{findKeywordInPCcluster}
\alias{findKeywordInPCcluster}
\title{Find the rank of your keyword in the PCcluster's GSEA annotation}
\usage{
findKeywordInPCcluster(PCAmodel, keyword, ind, n = NULL, includeTotal = FALSE)
}
\arguments{
\item{PCAmodel}{PCAGenomicSignatures-object.}

\item{keyword}{A character vector. If you are searching for multiple keywords
at the same time, use \code{\link{paste}} with \code{collapse = "|"} argument.}

\item{ind}{An integer. The PCcluster number you want to check.}

\item{n}{An interger. The number of top enriched pathways (based on abs(NES))
to search. Under default (\code{NULL}), all the enriched pathways are used.}

\item{includeTotal}{Default is \code{FALSE}. If it is set to \code{TRUE}, the
total number of enriched pathways will be also printed out as an output.}
}
\value{
A character containing the rank of keyword-containing pathways (separated
by |), followed by the total number of enriched pathways in parenthesis.
}
\description{
Once you provide PCAmodel, keyword you're searching for, and the PCcluster number
to this function, it will give you the abs(NES)-based rank of your keyword in
the enriched pathways of the target PCcluster. If can be useful to find out how
uniquely your keyword-containing pathways are represented.
}
