% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAvgLoading.R
\name{findStudiesInCluster}
\alias{findStudiesInCluster}
\title{Find the studies contributing each PCcluster}
\usage{
findStudiesInCluster(PCAmodel, ind = NULL, clustering = TRUE)
}
\arguments{
\item{PCAmodel}{PCAGenomicSignatures object.}

\item{ind}{A numeric vector containing the index of PCclusters you want to
find the related studies. Default is \code{NULL}.}

\item{clustering}{Under the default condition (\code{TRUE}), Kmeans clustering
will be done and you should provide \code{k}. If \code{FALSE}, you don't need
to provide \code{k}, but should provide pre-defined cluster membership of your
data through \code{cluster} argument.}
}
\value{
A list of character vector. Under the default conditoin (\code{ind = NULL}),
all the PCclusters will be checked for their contributing studies and the length
of the list will be same as the number of PCclusters (= \code{metadata(x)$k}).
}
\description{
Find the studies contributing each PCcluster
}
