% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicSignatures-class.R
\docType{class}
\name{GenomicSignatures-class}
\alias{GenomicSignatures-class}
\alias{GenomicSignatures}
\title{Construct \code{GenomicSignatures} object}
\value{
A \code{GenomicSignatures} object
}
\description{
GenomicSignatures
}
