% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcForGSEA_plot.R
\name{gseaNetwork}
\alias{gseaNetwork}
\title{Plot the network of enriched pathways}
\usage{
gseaNetwork(
  ind,
  PCAmodel,
  category = category,
  n = n,
  pvalueCutoff = pvalueCutoff,
  gseaRes = NULL,
  similarity_metric = c("jaccard_similarity", "overlap_similarity"),
  similarity_cutoff = 0.3,
  title = ""
)
}
\arguments{
\item{ind}{An interger. Index of avgLoading/PCcluster to apply GSEA.}

\item{PCAmodel}{PCAGenomicSignature object.}

\item{category}{A character vector representing MSigDB category. Options are
"H", "C1", "C2"(default), "C3", "C4", "C5", "C6", and "C7"}

\item{n}{An interger. The number of top and bottom enriched pathways to plot. Default is 10.}

\item{pvalueCutoff}{Cutoff for both pvalue and p.adjust. Default is 0.5.}

\item{gseaRes}{An output from \link{msigdb_gsea} function. If this argument
is provided, you don't need to provide the above inputs: ind, PCAmodel, category, n, pvalueCutoff.}

\item{similarity_metric}{Metric to calculate geneset similarity. Available values
are \code{c("jaccard_similarity", "overlap_similarity")}.}

\item{similarity_cutoff}{Geneset similarity cutoff. Default is 0.3.}

\item{title}{Plot title}
}
\value{
A visNetwork object
}
\description{
Plot the network of enriched pathways
}
\note{
Modified from \code{\link[hypeR]{hyp_emap}}
}
