% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapTable.R
\name{heatmapTable}
\alias{heatmapTable}
\title{Validation result in heatmap format}
\usage{
heatmapTable(
  data,
  row_title,
  num.out = 5,
  cutoff = NULL,
  breaks = c(0, 0.5, 1),
  colors = c("white", "white smoke", "red"),
  ...
)
}
\arguments{
\item{data}{An output matrix from \code{\link{validate}} function. Subset of
this matrix is plotted as a heatmap using \code{\link[ComplexHeatmap]{Heatmap}}}

\item{row_title}{A character string. Provide the row title.}

\item{num.out}{A number of highly validated PCclusters to subset to. Default is 5.
If \code{cutoff} is provided, this argument is ignored.}

\item{cutoff}{A numeric value for the minimum correlation. If you don't specify
this argument, this function will output the number of highly validated PCclusters
defined through \code{num.out}.}

\item{breaks}{A numeric vector of length 3. Number represents the values assigned
to three colors. Default is \code{c(0, 0.5, 1)}.}

\item{colors}{A character vector of length 3. Each represents the color assigned
to three breaks. Default is \code{c("white", "white smoke", "red")}.}

\item{...}{any additional argument for \code{ComplexHeatmap::Heatmap}}
}
\value{
A heatmap with the validation result subset through the given conditions.
}
\description{
Validation result in heatmap format
}
