% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawWordcloud.R
\name{meshTable}
\alias{meshTable}
\title{Build a two-column word/frequency table}
\usage{
meshTable(PCAmodel, ind, rm.noise = 4, weighted = TRUE)
}
\arguments{
\item{PCAmodel}{A PCAGenomicSignatures object}

\item{ind}{An index of PCcluster}

\item{rm.noise}{An integer. Any MeSH term found less than the given value here
will be excluded from wordcloud. If \code{rm.noise = 0}, all the MeSH terms
in PCcluster will be used. Default is 4.}

\item{weighted}{A logical. If \code{TRUE}, MeSH terms from each study are
weighted based on the variance explained by the principle component of the
study contributing a give PCcluster. Default is \code{TRUE}.}
}
\value{
A table with two columns, \code{word} and \code{freq}. MeSH terms in
the defined PCcluster (by \code{ind} argument) is ordered based on their frequency.
}
\description{
Build a two-column word/frequency table
}
