% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miniAllZ}
\alias{miniAllZ}
\title{Subset of allZ matrix constructed from 8 CRC training datasets}
\format{
A matrix with 100 genes and 160 PCs from 8 training datasets.
}
\source{
https://github.com/shbrief/model_building/tree/main/RAVmodel_8CRC
}
\usage{
miniAllZ
}
\description{
Eight colorectal cancer microarray datasets were used to build RAVmodel and
the intermediate file containing genes and top PCs from each dataset is named
as \code{allZ}. Hierarchical clustering result of \code{allZ} is saved as
\code{res_hcut}. For demonstration, we subset the \code{allZ} matrix with the
first 100 genes, which is named as \code{miniAllZ}.
}
\author{
Sehyun Oh \email{shbrief@gmail.com}
}
\keyword{data}
