% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miniPCAmodel}
\alias{miniPCAmodel}
\title{PCAGenomicSignatures from 536 studies annotated with MSigDB C2}
\format{
An object of class \code{PCAGenomicSignatures} with 13934 rows and 17 columns.
}
\usage{
miniPCAmodel
}
\description{
A object providing a reduced version of PCAmodel_C2 (PCAGenomicSignatures
object constructed from 536 studies and annotated with MSigDB C2).
}
\examples{
## The following is the code used to create this mini PCAmodel from the full
## PCAmodel_PLIERpriors.The full PCAmodel_PLIERpriors was created by the pipeline
## at https://github.com/shbrief/model_building.
\dontrun{
  library(PCAGenomicSignatures)
  library(bcellViper)
  data(bcellViper)
  getModel("PLIERpriors")
  PCAmodel <- readRDS("PCAmodel_PLIERpriors.rds")

  # include top 15 validated RAVs
  val_all <- validate(dset, PCAmodel)
  ind_validated <- validatedSignatures(val_all, num.out = 15, indexOnly = TRUE)

  # include 2 keyword-containing RAVs
  ind_keyword <- findSignature(PCAmodel, "Bcell", k = 5)

  ind_all <- unique(c(ind_validated, ind_keyword))
  miniPCAmodel <- PCAmodel[,ind_all]
  metadata(miniPCAmodel)$size <- metadata(miniPCAmodel)$size[ind_all]

  save(miniPCAmodel, file = "data/miniPCAmodel.RData", compress = "bzip2")
}
}
\keyword{data}
