% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miniRAVmodel}
\alias{miniRAVmodel}
\title{RAVmodel from 536 studies, annotated with MSigDB C2}
\format{
An object of class \code{PCAGenomicSignatures} with 13934 rows and 17 columns.
}
\usage{
miniRAVmodel
}
\description{
A object providing a miniature version of RAVmodel_C2 (PCAGenomicSignatures
object constructed from 536 studies and annotated with MSigDB C2).
}
\keyword{data}
