% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcForGSEA_plot.R
\name{msigdb_gsea}
\alias{msigdb_gsea}
\title{Barplot GSEA output}
\usage{
msigdb_gsea(ind, PCAmodel, category = "C2", n = 10, pvalueCutoff = 0.5)
}
\arguments{
\item{ind}{An interger. Index of avgLoading/PCcluster to apply GSEA.}

\item{PCAmodel}{PCAGenomicSignature object.}

\item{category}{A character vector representing MSigDB category. Options are
"H","C1","C2"(default), "C3","C4","C5","C6", and "C7"}

\item{n}{An interger. The number of top and bottom enriched pathways to plot. Default is 10.}

\item{pvalueCutoff}{Cutoff for both pvalue and p.adjust. Default is 0.5.}
}
\value{
Barplot of GSEA output. Top and bottom \code{n} genesets based on NES
are plotted and qvalues are denoted by color.
}
\description{
Barplot GSEA output
}
