% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAGenomicSignatures-class.R
\docType{class}
\name{PCAGenomicSignatures-class}
\alias{PCAGenomicSignatures-class}
\alias{.PCAGenomicSignatures}
\title{PCAGenomicSignatures}
\description{
PCAGenomicSignatures
}
\section{Slots}{

\describe{
\item{\code{trainingData}}{Containing metadata of each studies used in model building}
}}

