% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAGenomicSignatures-class.R
\name{PCAGenomicSignatures}
\alias{PCAGenomicSignatures}
\title{Construct \code{PCAGenomicSignatures} object}
\usage{
PCAGenomicSignatures(...)
}
\arguments{
\item{...}{Additional arguments for supporting functions}
}
\description{
The default contents of \code{PCAGenomicSignatures} object, with a set
of accessor and setter generic functions. When you create this object, \code{colData(x)$studies}
should be populated before adding any information in \code{trainingData(x)} slot
\itemize{
    \item assay(x) : PCAmodel (= avgLoading) containing genes x PCclusters
    \item metadata(x)$cluster : A vector of integers (from 1:k) indicating the
    cluster to which each point is allocated. Check \code{stats::kmeans} function.
    \item metadata(x)$size : The number of points in each cluster. Check \code{stats::kmeans} function.
    \item metadata(x)$iter : The number of (outer) iterations. Check \code{stats::kmeans} function.
    \item metadata(x)$k : The number of PCclusters.
    \item metadata(x)$n : The number of top PCs from each dataset.
    \item metadata(x)$geneSets : Name of the prior gene sets used to annotate average loadings.
    \item colData(x)$studies : A list of character vectors containing studies contributing to each PC cluster.
    \item colData(x)$gsea : A list of \code{gseaResult} objects. Build using \code{clusterProfiler::GSEA} function.
}
}
