% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawWordcloud.R
\name{PCinCluster}
\alias{PCinCluster}
\title{Extract the list of PCs in a cluster}
\usage{
PCinCluster(x, ind)
}
\arguments{
\item{x}{A PCAGenomicSignatures object}

\item{ind}{An index of PCcluster}
}
\description{
Extract the list of PCs in a cluster
}
