% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawWordcloud.R
\name{PCinRAV}
\alias{PCinRAV}
\title{Extract the list of PCs in a cluster}
\usage{
PCinRAV(PCAmodel, ind)
}
\arguments{
\item{PCAmodel}{A PCAGenomicSignatures object}

\item{ind}{An index of RAV}
}
\description{
Extract the list of PCs in a cluster
}
