% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotValidate.R
\name{plotValidate}
\alias{plotValidate}
\title{Plot validation results in an interactive graph}
\usage{
plotValidate(
  res,
  minClusterSize = 1,
  swFilter = FALSE,
  minSilhouetteWidth = 0,
  interactive = FALSE,
  colorPalette = "Dark2"
)
}
\arguments{
\item{res}{Output from \code{\link{validate}} function.}

\item{minClusterSize}{The minimum size of clusters to exclude from plotting.
Default value is 1, so any single-element clusters are excluded.}

\item{swFilter}{If \code{swFilter=TRUE}, only PCcluster above the cutoff, defined
through \code{minSilhouetteWidth} argument will be plotted. Default is \code{swFilter=FALSE}}

\item{minSilhouetteWidth}{A minimum average silhouette width to be plotted. Only
effective under \code{swFilter=TRUE} condition. Default is 0.}

\item{interactive}{If set to \code{TRUE}, the output will be interactive plot.
Default is \code{FALSE}.}

\item{colorPalette}{Default is \code{Dark2}. For other color options, please
check \code{\link[ggplot2]{scale_color_brewer}}}
}
\description{
There are three main information on the graph:
\itemize{
    \item x-axis : Pearson correlation coefficient. Higher value means that
    test dataset and PCcluster is more tightly associated with.
    \item y-axis : Silhouette width representing the quality of PCclusters.
    \item size : The number of studies in each PCcluster. (= cluster size)
    \item color : Test dataset's PC number that validate each PCcluster. Because we
    used top 8 PCs of the test dataset, there are 8 categories.
}
}
