% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAGenomicSignatures-methods.R
\name{RAVName}
\alias{RAVName}
\alias{.RAVName}
\title{Formatting RAV name}
\usage{
.RAVName(x, ...)
}
\arguments{
\item{x}{PCAGenomicSignature-object}

\item{...}{Additional arguments for supporting functions.}
}
\description{
Keep the name with 'k + cluster number + number of PCs + number of unique studies'
info during the model construction to make it easy to keep track of them, but at the
PCAGenomicSignatures-class object building step, covert them into 'RAV + cluster number'.
}
