% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_hcut}
\alias{res_hcut}
\title{Subset of allZ matrix constructed from 8 CRC training datasets}
\format{
\code{hclust} object from \code{factoextra::hcut} function.
}
\usage{
res_hcut
}
\description{
Eight colorectal cancer microarray datasets were used to build RAVmodel and
the intermediate file containing genes and top PCs from each dataset is named
as \code{allZ}. Hierarchical clustering result of \code{allZ} is saved as
\code{res_hcut}.
}
\author{
Sehyun Oh \email{shbrief@gmail.com}
}
\keyword{data}
