% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmNaInf.R
\name{rmNaInf}
\alias{rmNaInf}
\title{Remove rows with missing and Inf values from a matrix}
\usage{
rmNaInf(x)
}
\arguments{
\item{x}{A numeric matrix.}
}
\value{
The updated input matrix where rows with
   NA and Inf values are removed.
}
\description{
Remove rows with missing and Inf values from a matrix
}
\examples{
m = matrix(rnorm(100),ncol=10)
m[1,1] = NA

m1 = rmNaInf(m)
dim(m1)

}
