% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowNorm.R
\name{rowNorm}
\alias{rowNorm}
\title{z-score each row of the data}
\usage{
rowNorm(x)
}
\arguments{
\item{x}{A gene-expression matrix with genes in rows and samples in columns}
}
\value{
a matrix with each row mean centered and scaled by rowwise sd
}
\description{
z-score each row of the data
}
\examples{
x = matrix(rnorm(100),nc=10)
y = rowNorm(x)
apply(y,1,mean)

}
