% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcForGSEA.R
\name{run_gsea}
\alias{run_gsea}
\title{GSEA on pre-ordered gene lists}
\usage{
run_gsea(
  geneList,
  TERM2GENE,
  TERM2NAME,
  minGSSize = 10,
  maxGSSize = 500,
  pvalueCutoff = 0.05,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{geneList}{A list of genes ordered by rank}

\item{TERM2GENE}{User input annotation of TERM TO GENE mapping, a data frame
of 2 column with term and gene}

\item{TERM2NAME}{User input of TERM TO NAME mapping, a data.frame of 2 column
with term and name. Optional.}

\item{minGSSize}{A mininum size of gene set to be analyzed}

\item{maxGSSize}{A maximum size of gene set to be analyzed}

\item{pvalueCutoff}{p-value cutoff}

\item{verbose}{Logical. Default is \code{FALSE}}

\item{...}{Any additional argument inherited from
\code{\link[clusterProfiler]{GSEA}}.}
}
\value{
A list of \code{gseaResult} objects. If there is no enrichment
result, \code{NA} will be returned.
}
\description{
This function is a wrapper of
\code{\link[clusterProfiler]{GSEA}} function, making it applicable to a list
of gene lists. Set seed for reproducible result.
}
