% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcForGSEA.R
\name{subsetEnrichedPathways}
\alias{subsetEnrichedPathways}
\title{Subset enriched pathways of RAV}
\usage{
subsetEnrichedPathways(PCAmodel, ind = NULL, n = 10, both = FALSE)
}
\arguments{
\item{PCAmodel}{PCAGenomicSignatures object. Also an output from \code{\link[clusterProfiler]{GSEA}} can be used.}

\item{ind}{A numeric vector containing the PCcluster number you want to check
enriched pathways. If not specified, this function returns results from all the PCclusters.}

\item{n}{The number of top and bottom pathways to be selected based on normalized
enrichment score (NES).}

\item{both}{Default is \code{FALSE}, where only the top \code{n} pathways will
be printed. If it is set to \code{TRUE}, the ouput will contain both top and
bottom \code{n} pathways.}
}
\value{
A data frame with top and bottom \code{n} pathways from the enrichment results.
}
\description{
This function finds the top enriched pathways of a given RAV. This function is
renamed from \code{subsetPathways} to \code{subsetEnrichedPathways}.
}
