% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcForGSEA.R
\name{subsetGSEA}
\alias{subsetGSEA}
\title{Subset GSEA output}
\usage{
subsetGSEA(gseaRes, n = 20)
}
\arguments{
\item{gseaRes}{An output from \code{\link[clusterProfiler]{GSEA}}}

\item{n}{A number of output to keep based on the \code{abs(NES)}}
}
\description{
Subset GSEA output
}
