% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapTable.R
\name{validatedIndex}
\alias{validatedIndex}
\title{Validation result in data frame}
\usage{
validatedIndex(data, num.out = 5, cutoff = NULL)
}
\arguments{
\item{data}{An output matrix from \code{\link{validate}} function.}

\item{num.out}{A number of highly validated PCclusters to output. Default is 5.
If \code{cutoff} is provided, this argument is ignored.}

\item{cutoff}{A numeric value for the minimum correlation. If you don't specify
this argument, this function will output the number of highly validated PCclusters
defined through \code{num.out}.}
}
\value{
A subset of the input matrix, which meets the given condition.
}
\description{
Validation result in data frame
}
