% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePCcluster.R
\name{annotateRAV}
\alias{annotateRAV}
\title{Search the top enriched pathways for RAV}
\usage{
annotateRAV(PCAmodel, ind, n = 5, abs = FALSE)
}
\arguments{
\item{PCAmodel}{PCAGenomicSignatures object.}

\item{ind}{An integer for RAV you want to check the enriched pathways.}

\item{n}{A number of top enriched pathways to output. Default is 5.}

\item{abs}{Default is \code{FALSE}. If it's set to \code{TRUE}, the enriched
pathways will be listed based on \code{abs(NES)}.}
}
\value{
A data frame with \code{n} rows and 4 columns; Description, NES, pvalue, and qval
}
\description{
Search the top enriched pathways for RAV
}
\examples{
data(miniPCAmodel)
library(bcellViper)
data(bcellViper)
annotateRAV(miniPCAmodel, ind = 695)
#                 Description      NES pvalue      qvalues
# 1 IRIS_Bcell-Memory_IgG_IgA 2.763472  1e-10 5.824561e-09
# 2              DMAP_BCELLA3 2.726775  1e-10 5.824561e-09
# 3     IRIS_Bcell-Memory_IgM 2.701570  1e-10 5.824561e-09
# 4          IRIS_Bcell-naive 2.699145  1e-10 5.824561e-09
# 5              DMAP_BCELLA4 2.698821  1e-10 5.824561e-09

}
