% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonGene.R
\name{commonGene}
\alias{commonGene}
\title{Common genes from datasets}
\usage{
commonGene(dat, loaded = TRUE)
}
\arguments{
\item{dat}{A character vector with name of the datasets loaded in memory. You
can also provide a list of genes you want to merge into a common gene list.}

\item{loaded}{Under the default condition (\code{TRUE}), the dataset (e.g. expression
matrix with genes as row names) is loaded. \code{FALSE} if you provide a list
of gene in the argument \code{dat}.}
}
\value{
A character vector with the common genes of the datasets.
}
\description{
Common genes from datasets
}
