% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.loadingCor}
\alias{.loadingCor}
\title{Validating new dataset}
\usage{
.loadingCor(dataset, avgLoading, method = "pearson", scale = FALSE)
}
\arguments{
\item{dataset}{A expression dataset to validate. Genes in rows and samples in
columns. Gene names should be in 'symbol' format. It can be ExpressionSet,
SummarizedExperiment, RangedSummarizedExperiment, or matrix.}

\item{avgLoading}{A matrix with genes by RAVs.}

\item{method}{A character string indicating which correlation coefficient is
to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{scale}{Default is \code{FALSE}. If it is set to \code{TRUE}, dataset will
be row normalized by \link{rowNorm} function.}
}
\value{
A matrix of Pearson correlation coefficient (default, defined through \code{method}
argument) between RAVs (row) and the top 8 PCs from the datasets (column)
}
\description{
Validating new dataset
}
