% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAvgLoading.R
\name{findStudiesInCluster}
\alias{findStudiesInCluster}
\title{Find the studies contributing each PCcluster}
\usage{
findStudiesInCluster(PCAmodel, ind = NULL, studyTitle = FALSE)
}
\arguments{
\item{PCAmodel}{PCAGenomicSignatures object.}

\item{ind}{A numeric vector containing the index of PCclusters you want to
find the related studies. Default is \code{NULL}.}

\item{studyTitle}{Default is \code{FALSE}. This parameter is effective only when
the \code{index} value is specificed. If it's \code{TRUE}, the output will be
a data frame with the study}
}
\value{
A list of character vector. Under the default conditoin (\code{ind = NULL}),
all the PCclusters will be checked for their contributing studies and the length
of the list will be same as the number of PCclusters (= \code{metadata(x)$k}).
}
\description{
Find the studies contributing each PCcluster
}
\note{
Mainly used for modeling building, within \link{buildAvgLoading}.
}
