% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicSignatures-class.R
\docType{class}
\name{GenomicSignatures-class}
\alias{GenomicSignatures-class}
\title{Virtual class inherited from SummarizedExperiment}
\arguments{
\item{x}{A \code{GenomicSignatures-class} object}

\item{value}{See details.}
}
\description{
GenomicSignatures is a virtual class inherited from
SummarizedExperiment and hosts GenomicSignatures models built from different
dimensional reduction methods. Currently, PCA-based model, called
PCAGenomicSignatures, is available.
}
\details{
GenomicSignatures
}
