% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModel.R
\name{getModel}
\alias{getModel}
\title{Download the PCAmodel from GCP}
\usage{
getModel(prior = c("C2", "PLIERpriors"), dir)
}
\arguments{
\item{prior}{The name of gene sets used to annotate PCAGenomicSignatures. Currently
there are two available options.
\itemize{
    \item \code{C2} : MSigDB C2 (curated gene sets)
    \item \code{PLIERpriors} : bloodCellMarkersIRISDMAP, svmMarkers, and canonicalPathways
}}

\item{dir}{A path to the directory where PCAGenomicSginatures object will be downloaded.}
}
\value{
PCAGenomicSignatures object in \code{.rds} format saved in the user
defined location.
}
\description{
Download the PCAmodel from GCP
}
