% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModel.R
\name{getModel}
\alias{getModel}
\title{Access a PCAGenomicSignature model}
\usage{
getModel(prior = c("C2", "PLIERpriors"))
}
\arguments{
\item{prior}{The name of gene sets used to annotate PCAGenomicSignatures. Currently
there are two available options.
\itemize{
    \item \code{C2} : MSigDB C2 (curated gene sets)
    \item \code{PLIERpriors} : bloodCellMarkersIRISDMAP, svmMarkers, and canonicalPathways
}}
}
\value{
PCAGenomicSignatures object in \code{.rds} format saved in the user
defined location.
}
\description{
Access a PCAGenomicSignature model
}
