% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModel.R
\name{getModel}
\alias{getModel}
\title{Download a PCAGenomicSignatures model}
\usage{
getModel(prior = c("C2", "PLIERpriors"), load = FALSE)
}
\arguments{
\item{prior}{The name of gene sets used to annotate PCAGenomicSignatures. Currently
there are two available options.
\itemize{
    \item \code{C2} : MSigDB C2 (curated gene sets)
    \item \code{PLIERpriors} : bloodCellMarkersIRISDMAP, svmMarkers, and canonicalPathways
}}

\item{load}{Default is \code{FALSE}. If it's set to \code{TRUE}, the downloaded
file will be loaded.}
}
\value{
File cache location or PCAGenomicSignatures object loaded from it.
}
\description{
Download a PCAGenomicSignatures model
}
