% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapTable.R
\name{heatmapTable}
\alias{heatmapTable}
\title{Validation result in heatmap format}
\usage{
heatmapTable(
  val_all,
  ind = NULL,
  num.out = 5,
  scoreCutoff = NULL,
  swCutoff = NULL,
  clsizeCutoff = NULL,
  breaks = c(0, 0.5, 1),
  colors = c("white", "white smoke", "red"),
  column_title = NULL,
  row_title = NULL,
  whichPC = NULL,
  ...
)
}
\arguments{
\item{val_all}{An output matrix from \code{\link{validate}} function. Subset of
this matrix is plotted as a heatmap using \code{\link[ComplexHeatmap]{Heatmap}}}

\item{ind}{An integer vector. If this parameter is provided, the other parameters,
\code{num.out, scoreCutoff, swCutoff, clsizeCutoff} will be ignored and the heatmap
table containing only the provided index will be printed.}

\item{num.out}{A number of highly validated PCclusters to output. Default is 5.
If any of the cutoff parameters are provided, \code{num.out} or the number of
filtered PCclusters, whichever smaller, will be chosen.}

\item{scoreCutoff}{A numeric value for the minimum correlation.}

\item{swCutoff}{A numeric value for the minimum average silhouette width.}

\item{clsizeCutoff}{A integer value for the minimum cluster size.}

\item{breaks}{A numeric vector of length 3. Number represents the values assigned
to three colors. Default is \code{c(0, 0.5, 1)}.}

\item{colors}{A character vector of length 3. Each represents the color assigned
to three breaks. Default is \code{c("white", "white smoke", "red")}.}

\item{column_title}{A character string. Provide the column title.}

\item{row_title}{A character string. Provide the row title.}

\item{whichPC}{An integer value between 1 and 8. PC number of your data to check
the validated signatures with. Under the default (\code{NULL}), it outputs top
scored signatures with any PC of your data.}

\item{...}{any additional argument for \code{\link[ComplexHeatmap]{Heatmap}}}
}
\value{
A heatmap with the validation result subset through the given conditions.
}
\description{
Validation result in heatmap format
}
