% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miniRAVmodel}
\alias{miniRAVmodel}
\title{RAVmodel from 536 studies, annotated with MSigDB C2}
\format{
PCAGenomicSignatures
}
\usage{
miniRAVmodel
}
\description{
A object providing a miniature version of RAVmodel_C2 (PCAGenomicSignatures
object constructed from 536 studies and annotated with MSigDB C2).
}
\author{
Sehyun Oh \email{shbrief@gmail.com}
}
\keyword{data}
