% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcForGSEA.R
\name{msigdb_gsea}
\alias{msigdb_gsea}
\title{MSigDB GSEA results}
\usage{
msigdb_gsea(
  ind,
  RAVmodel,
  category = "C2",
  n = NULL,
  pvalueCutoff = 0.5,
  minGSSize = 10,
  maxGSSize = 500,
  pAdjustMethod = "BH",
  verbose = FALSE,
  seed = FALSE,
  by = "fgsea",
  geneSets = NULL
)
}
\arguments{
\item{ind}{An interger. Index of RAV to apply GSEA.}

\item{RAVmodel}{PCAGenomicSignature object.}

\item{category}{A character vector representing MSigDB category. Options are
"H", "C1", "C2"(default), "C3", "C4", "C5", "C6", and "C7"}

\item{n}{An interger. The number of top and bottom enriched pathways to plot.
Default is \code{NULL} and print out all the pathways enriched under \code{pvalueCutoff}.}

\item{pvalueCutoff}{Cutoff for both pvalue and p.adjust. Default is 0.5.}

\item{minGSSize}{A mininum size of gene set to be analyzed}

\item{maxGSSize}{A maximum size of gene set to be analyzed}

\item{pAdjustMethod}{p-value adjustment methods, which will be used as an input
for \code{method} argument of \code{\link[stats]{p.adjust}} function. Available
options are "holm", "hochberg", "hommel", "bonferroni", "BH"(default), "BY", "fdr", "none".}

\item{verbose}{Logical. Default is FALSE.}

\item{seed}{Logical. Default is FALSE.}

\item{by}{Available options are \code{c("fgsea", "DOSE")}. Default is "fgsea".}

\item{geneSets}{Custom genesets to use with MSigDB genesets. It should be in
a named list format.}
}
\value{
Barplot of GSEA output. Top and bottom \code{n} genesets based on NES
are plotted and qvalues are denoted by color.
}
\description{
MSigDB GSEA results
}
