% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicSignatures-class.R
\docType{class}
\name{PCAGenomicSignatures-class}
\alias{PCAGenomicSignatures-class}
\alias{.PCAGenomicSignatures}
\title{PCAGenomicSignatures-class}
\arguments{
\item{x}{A \code{GenomicSignatures-class} object}

\item{value}{See details.}
}
\description{
PCA-based \code{\link{GenomicSignatures}}.
}
\details{
PCAGenomicSignatures
}
\section{Slots}{

\describe{
\item{\code{trainingData}}{A \code{\link[S4Vectors]{DataFrame}} class object for metadata
associated with training data}
}}

