% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAGenomicSignatures-class.R
\name{PCAGenomicSignatures-methods}
\alias{PCAGenomicSignatures-methods}
\title{Accessing and modifying information in PCAGenomicSignatures}
\arguments{
\item{x}{A \code{PCAGenomicSignatures} object}

\item{value}{See details.}
}
\description{
A set of accessor and setter generic functions to extract
either the \code{assay}, \code{colData}, or \code{metadata} slots of a
\code{\link{PCAGenomicSignatures}} object
}
\section{Setters}{

Setter method values (i.e., \code{function(x) <- value}):
\itemize{
    \item geneSets<- : A character vector containing the name of gene sets used
    to annotate average loadings
    \item studies<- : A list of character vectors containing gene sets used to annotate average loadings
    \item gsea<- : A list of \code{gseaResult} objects.
    \item metadata<- : A \code{list} object of metadata
    \item `$<-` : A vector to replace the indicated column in \code{colData}
}
}

\section{Accessors}{

All the accessors inherited from \code{SummarizedExperiment} are available and
the additional accessors for \code{PCAGenomicSignatures} specific data are listed
below.
\itemize{
   \item model : Equivalent to the \code{assays(x)$model} accessor for convenience
   \item geneSets : Access the \code{metadata(x)$geneSets} slot
   \item studies : Access the \code{colData(x)$studies} slot
   \item gsea : Access the \code{colData(x)$gsea}
   \item `$` : Access a column in \code{colData}
   \item trainingData : Access the \code{trainingData} slot
   \item mesh : Access the \code{trainingData(x)$MeSH} slot
   \item PCAsummary : Access the \code{trainingData(x)$PCAsummary} slot
}
}

