% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawWordcloud.R
\name{PCinRAV}
\alias{PCinRAV}
\title{Extract the list of PCs in a cluster}
\usage{
PCinRAV(RAVmodel, ind)
}
\arguments{
\item{RAVmodel}{A PCAGenomicSignatures object}

\item{ind}{An index of RAV}
}
\value{
A character vector of PC/study names
}
\description{
A RAV model contain clusters of PCs from individual
studies. This function extracts the names of the original
PCs from the RAV model given the index in the RAV model.
}
\examples{
data(miniRAVmodel)
PCinRAV(miniRAVmodel,695)

}
