% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnnotedPCA.R
\name{plotAnnotatedPCA}
\alias{plotAnnotatedPCA}
\title{Two-dimentional PCA plot with the PC annotation}
\usage{
plotAnnotatedPCA(
  dataset,
  PCAmodel,
  PCs,
  val_all = NULL,
  color_by = NULL,
  color_lab = NULL,
  trimed_pathway_len = 45
)
}
\arguments{
\item{dataset}{An expression matrix with genes (rows) x samples (columns)}

\item{PCAmodel}{PCAGenomicSignatures-class object}

\item{PCs}{A numeric vector length of 2. It should be between 1 and 8.}

\item{val_all}{The output from \code{\link{validate}}}

\item{color_by}{A named vector with the feature you want to color by. Name should
be match with the sample names of the dataset.}

\item{color_lab}{A name for color legend. If this argument is not provided, the
color legend will be labeld as "Color By" by default.}

\item{trimed_pathway_len}{Positive inter values, which is the display width of
pathway names. Default is 45.}
}
\description{
Two-dimentional PCA plot with the PC annotation
}
