% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmNaInf.R
\name{rmNaInf}
\alias{rmNaInf}
\title{Remove missing and Inf values from expression dataset}
\usage{
rmNaInf(x)
}
\arguments{
\item{x}{A matrix with the expression values.}
}
\value{
The updated input matrix where NA and Inf values are removed.
}
\description{
Remove missing and Inf values from expression dataset
}
