% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowNorm.R
\name{rowNorm}
\alias{rowNorm}
\title{z-score each row of the data}
\usage{
rowNorm(x)
}
\arguments{
\item{x}{A gene-expression matrix with genes in rows and samples in columns}
}
\description{
z-score each row of the data
}
\author{
\url{https://github.com/shbrief/PLIER/blob/master/R/Allfuncs.R}
}
