% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatedSignatures.R
\name{validatedSignatures}
\alias{validatedSignatures}
\title{Validation result in data frame}
\usage{
validatedSignatures(
  data,
  num.out = 5,
  scoreCutoff = NULL,
  swCutoff = NULL,
  clsizeCutoff = NULL
)
}
\arguments{
\item{data}{An output matrix from \code{\link{validate}} function.}

\item{num.out}{A number of highly validated PCclusters to output. Default is 5.
If any of the cutoff parameters are provided, \code{num.out} or the number of
filtered PCclusters, whichever smaller, will be chosen.}

\item{scoreCutoff}{A numeric value for the minimum correlation.}

\item{swCutoff}{A numeric value for the minimum average silhouette width.}

\item{clsizeCuoff}{A integer value for the minimum cluster size.}
}
\value{
A subset of the input matrix, which meets the given condition.
}
\description{
Validation result in data frame
}
